refactor(copyright): extract token_utils sub-module from detector.rs
Move token collection, filtering, and detection builder functions
(970 lines) into detector_token_utils.rs following the existing
#[path] sub-module pattern. Re-export via pub(super) use.
